agents/agents/metadata-extractor-agent/requirements.txt
uagents>=0.12.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
